requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
torch
groq
pydantic
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
